Make sure regular expressions for long operators come before prefixes
The lexer in this project operates in a very simple way: it tries each
regular expression in a list until one matches. Previously, the
regex for "<" came before "<=". Since "<" is a prefix of "<=", it
would get matched, and we would end up with an error. This change
fixes the order of the regexes.